Rename namespace to 'Codidact.Authentication.WebApp'.
