+++ .vscode/launch.json
@@ -6,13 +6,10 @@
             "type": "coreclr",
             "request": "launch",
             "preLaunchTask": "build",
-            "program": "${workspaceFolder}/src/Web/bin/Debug/netcoreapp3.1/Codidact.Authentication.Web.dll",
-            "cwd": "${workspaceFolder}/src/Web",
+            "program": "${workspaceFolder}/src/WebApp/bin/Debug/netcoreapp3.1/Codidact.Authentication.WebApp.dll",
+            "cwd": "${workspaceFolder}/src/WebApp",
             "env": {
                 "ASPNETCORE_ENVIRONMENT": "Development",
-            },
-            "sourceFileMap": {
-                "/Views": "${workspaceFolder}/Views",
             },
             "logging": {
                 "moduleLoad": false,
