--- Codidact.Authentication.sln
@@ -1,84 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{189DA739-3474-42A2-9391-D025A56CF719}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "src\Domain\Domain.csproj", "{944BF95B-70C3-4610-A28A-6E85674C2FE9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "src\Application\Application.csproj", "{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "src\Infrastructure\Infrastructure.csproj", "{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Web", "src\Web\Web.csproj", "{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Debug|x64.Build.0 = Debug|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Debug|x86.Build.0 = Debug|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Release|x64.ActiveCfg = Release|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Release|x64.Build.0 = Release|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Release|x86.ActiveCfg = Release|Any CPU
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9}.Release|x86.Build.0 = Release|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Debug|x64.Build.0 = Debug|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Debug|x86.Build.0 = Debug|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Release|x64.ActiveCfg = Release|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Release|x64.Build.0 = Release|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Release|x86.ActiveCfg = Release|Any CPU
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215}.Release|x86.Build.0 = Release|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Debug|x64.Build.0 = Debug|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Debug|x86.Build.0 = Debug|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Release|x64.ActiveCfg = Release|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Release|x64.Build.0 = Release|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Release|x86.ActiveCfg = Release|Any CPU
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29}.Release|x86.Build.0 = Release|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Debug|x64.Build.0 = Debug|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Debug|x86.Build.0 = Debug|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Release|x64.ActiveCfg = Release|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Release|x64.Build.0 = Release|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Release|x86.ActiveCfg = Release|Any CPU
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{944BF95B-70C3-4610-A28A-6E85674C2FE9} = {189DA739-3474-42A2-9391-D025A56CF719}
-		{D134FA54-A157-4C90-9B4F-8ED5BFAC6215} = {189DA739-3474-42A2-9391-D025A56CF719}
-		{7C6E9851-DA68-4222-8CA4-EDF217B2CA29} = {189DA739-3474-42A2-9391-D025A56CF719}
-		{6C756367-ACC4-4A23-B1BE-B40A3EE6F170} = {189DA739-3474-42A2-9391-D025A56CF719}
-	EndGlobalSection
-EndGlobal
