+++ src/WebApp/Pages/_ViewImports.cshtml
@@ -1,2 +1,2 @@
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@using Codidact.Authentication.Web.Pages
+@namespace Codidact.Authentication.WebApp.Pages
