+++ src/WebApp/WebApp.csproj
@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
     <PropertyGroup>
         <TargetFramework>netcoreapp3.1</TargetFramework>
-        <RootNamespace>Codidact.Authentication.Web</RootNamespace>
-        <AssemblyName>Codidact.Authentication.Web</AssemblyName>
+        <RootNamespace>Codidact.Authentication.WebApp</RootNamespace>
+        <AssemblyName>Codidact.Authentication.WebApp</AssemblyName>
     </PropertyGroup>
 
     <ItemGroup>
